Editor Pane
*worked on editor pane and added class that manages it 
*Need to integrate it more cleanly into the tabview and add functionality

+++ src/DSMData/DSMItem.java
@@ -42,7 +42,7 @@ public class DSMItem {
         this.name = name;
     }
 
-    public void setSortIndex(int index) {
+    public void setSortIndex(double index) {
         this.sortIndex = index;
     }
